requirements.txt
# aplicação front/db
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.2.0
marshmallow==3.21.3
flask-smorest==0.44.0

# qualidade e testes
pytest==8.2.2
pytest-flask==1.3.0

# nucleo projeto audio features
numba==0.62.1
llvmlite==0.45.1
pandas==2.2.3
numpy==2.1.3
praat-parselmouth==0.4.3
librosa==0.10.2.post1
python-dateutil==2.9.0.post0
tqdm>=4.65.0
scipy>=1.10.0
matplotlib